+++ generator/julia.h
@@ -1,6 +1,6 @@
 #ifndef H_JULIA
 #define H_JULIA
 
-unsigned long julia(double lx, double ly, unsigned long max_iter);
+int julia(double lx, double ly, int max_iter);
 
 #endif
