+++ generator/mandelbrot.h
@@ -1,6 +1,6 @@
 #ifndef H_MANDELBROT
 #define H_MANDELBROT
 
-unsigned long mandelbrot(double lx, double ly, unsigned long max_iter);
+int mandelbrot(double lx, double ly, int max_iter);
 
 #endif
